+++ src/lib.rs
@@ -1,5 +1,7 @@
 extern crate serde;
 extern crate serde_json;
+#[macro_use]
+extern crate serde_derive;
 
 use std::io::stdin;
 use std::error::Error;
